requirements.txt
massive>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
lxml>=4.9.0
html5lib>=1.1
polygon-api-client>=1.0.0
